backend/requirements.txt
Django==5.2.10
djangorestframework==3.16.1
pandas==2.3.3
numpy==2.2.6
gunicorn==23.0.0
whitenoise==6.8.2

